CHG0005
Get Api calls -> Product
Create Api Calls -> Inventory
