+++ inventory-service/src/main/java/com/apnidukan/Service/IInventoryService.java
@@ -1,0 +1,7 @@
+package com.apnidukan.Service;
+
+import com.apnidukan.dto.ProductDTO;
+public interface IInventoryService {
+
+    ProductDTO createProduct(ProductDTO productDTO);
+}
