+++ inventory-service/src/main/java/com/apnidukan/repository/InventoryRepository.java
@@ -1,0 +1,7 @@
+package com.apnidukan.repository;
+
+import com.apnidukan.entity.Product;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface InventoryRepository extends JpaRepository<Product, Long> {
+}
